refactor: move keyboard functions into header
